Started rewriting the move generation to better optimize it, and making better evaluation techniques for better move selection
