+++ OnGoing/simple-chess/AI.py
@@ -9,7 +9,7 @@ class AiMovement:
 
     def pick_move(self):
         search = Search(self.board, self.player)
-        self.move = search.start_search(4)
+        self.move = search.start_search(5)
 
     def play_move(self):
         move_invoke = Invoke(self.move, self.board)
